requirements.txt
alembic==0.9.8
blinker==1.4
coverage==4.5.1
Flask==0.12.2
Flask-Cors==3.0.4
Flask-Script==2.0.6
Flask-GraphQL==1.4.1
graphene-sqlalchemy==2.0.0
graphene==2.1
MarkupSafe==1.0
more-itertools==4.1.0
psycopg2-binary==2.7.4
py==1.5.3
pytest==3.5.0
PyJWT==1.6.4
nose==1.3.7
python-editor==1.0.3
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
virtualenv==15.2.0
Flask-Testing==0.7.1
flake8==3.5.0
coveralls
